+++ packages/flutter_tools/test/integration.shard/test_utils.dart
@@ -107,3 +107,57 @@ Future<void> pollForServiceExtensionValue<T>({
     ' attempts responded with \'$continuePollingValue\'.',
   );
 }
+
+
+Future<void> evaluateTrivialExpressions(FlutterTestDriver flutter) async {
+  ObjRef res;
+
+  res = await flutter.evaluateInFrame('"test"');
+  expectValueOfType(res, InstanceKind.kString, 'test');
+
+  res = await flutter.evaluateInFrame('1');
+  expectValueOfType(res, InstanceKind.kInt, 1.toString());
+
+  res = await flutter.evaluateInFrame('true');
+  expectValueOfType(res, InstanceKind.kBool, true.toString());
+}
+
+Future<void> evaluateComplexExpressions(FlutterTestDriver flutter) async {
+  final ObjRef res = await flutter.evaluateInFrame('new DateTime.now().year');
+  expectValueOfType(res, InstanceKind.kInt, DateTime.now().year.toString());
+}
+
+Future<void> evaluateHotReloadedExpresison(FlutterTestDriver flutter, int expected) async {
+  final ObjRef res = await flutter.evaluateInFrame('foo()');
+  expectValueOfType(res, InstanceKind.kInt, expected.toString());
+}
+
+Future<void> evaluateComplexReturningExpressions(FlutterTestDriver flutter) async {
+  final DateTime now = DateTime.now();
+  final ObjRef resp = await flutter.evaluateInFrame('new DateTime.now()');
+  expectInstanceOfClass(resp, 'DateTime');
+  // Ensure we got a reasonable approximation. The more accurate we try to
+  // make this, the more likely it'll fail due to differences in the time
+  // in the remote VM and the local VM at the time the code runs.
+  final InstanceRef res = await flutter.evaluate(resp.id, r'"$year-$month-$day"');
+  expectValue(res, '${now.year}-${now.month}-${now.day}');
+}
+
+void expectInstanceOfClass(ObjRef result, String name) {
+  expect(result,
+    const TypeMatcher<InstanceRef>()
+      .having((InstanceRef instance) => instance.classRef.name, 'resp.classRef.name', name));
+}
+
+void expectValueOfType(ObjRef result, String kind, String message) {
+  expect(result,
+    const TypeMatcher<InstanceRef>()
+      .having((InstanceRef instance) => instance.kind, 'kind', kind)
+      .having((InstanceRef instance) => instance.valueAsString, 'valueAsString', message));
+}
+
+void expectValue(ObjRef result, String message) {
+  expect(result,
+    const TypeMatcher<InstanceRef>()
+      .having((InstanceRef instance) => instance.valueAsString, 'valueAsString', message));
+}
